### 2016-09-28 #83 @pjotrp
I was thinking of asking the biopython members @peterjc, what do yo think? Heng we should reserve for something more performance related - that is unless @lh3 wants to have a go at JOSS?
