### 2016-09-28 #83 @peterjc
I've not used VCF so I think that might rule me out. Perhaps @chapmanb would be a better match?
