### 2016-09-30 1d88940
Fixing issues detected by Landscape.io
