+++ .landscape.yml
@@ -1,0 +1,2 @@
+ignore-paths:
+  - vcfpy/__init__.py  # unused imports
