### 2016-09-30 e960665
Serialization of records in writer
