### 2016-10-02 #83 @chapmanb
Thanks for the ping, happy to help with this.
