### 2016-10-04 #85 @whedon
[REVIEW]: VCFPy: a Python 3 library with good support for both reading and writing VCF
**Submitting author:** @holtgrewe (<a href="http://orcid.org/0000-0002-3051-1763">Manuel Holtgrewe</a>)
**Repository:** <a href="https://github.com/bihealth/vcfpy" target ="_blank">https://github.com/bihealth/vcfpy</a>
**Version:** v0.7.0
**Archive:** <a href="https://dx.doi.org/10.6084/m9.figshare.3863583.v2" target="_blank">https://dx.doi.org/10.6084/m9.figshare.3863583.v2</a>
