### 2016-10-04 21e1741
Update paper.md
