### 2016-10-14 #85 @chapmanb
Thank you for the work on VCFPy. This software is open source, well documented, tested and generally pleasurable to use and contribute to. It easily passes all of the JOSS requirements listed above.

After testing, I don't have any major feature suggestions as it does work nicely for reading/writing VCFs. This is a welcome contributions, as changing VCFs is tricky with the current set of Python VCF tools (PyVCF, cyvcf2, pysam.VariantFile) and only a subset of changes are possible.

I do have concerns centered around development and support. My biggest worry is that this further splinters the Python VCF parsing community, making it more difficult to support and improve the parsers. While it provides another option filling a missing set of functionality, I would have much rather seen this added to an existing Python project. The biggest issue for most people with Python-only parses like PyVCF and now VCFPy is their speed. You mention issues building off of PyVCF, but would it have been possible to contribute some of this work to a htslib-based parser like pysam's.

Thanks again for this contribution and hope this feedback is useful.
