### 2016-10-14 #85 @pjotrp
Thank you @chapmanb! Note to the author that I wrote a fast parser in Ruby by processing multi-core, see https://github.com/pjotrp/bioruby-vcf. Perhaps worth looking into.

@arfon I think we can accept the paper. Is there a @whedon command for that?
