### 2016-10-14 #85 @arfon
> @arfon I think we can accept the paper. Is there a @whedon command for that?

Not yet. I'll move it forward from here.

@holtgrewe at this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
