### 2016-10-25 #85 @holtgrewe
@arfon Here it is https://dx.doi.org/10.6084/m9.figshare.3863583.v2

@chapmanb I agree that improving another library would have been better but I found working with the PyVCF and pysamtools code to be harder than implementing everything from scratch and not having to be backwards compatible.
