### 2016-10-25 #85 @arfon
@chapmanb many thanks for the review here!

@holtgrewe your paper is now accepted into JOSS. Your DOI is http://dx.doi.org/10.21105/joss.00085 🎉 🚀 💥
